+++ Sources/KeyboardKit/Actions/KeyboardAction+Rows.swift
@@ -1,0 +1,53 @@
+//
+//  KeyboardAction+Rows.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2019-07-04.
+//  Copyright © 2021-2023 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+public extension KeyboardAction {
+    
+    /// This is an alias for a ``KeyboardAction`` array.
+    typealias Row = [KeyboardAction]
+    
+    /// This is an alias for a ``KeyboardAction.Row`` array.
+    typealias Rows = [KeyboardAction.Row]
+}
+
+public extension KeyboardAction.Row {
+    
+    /// Map a string array to a keyboard action row.
+    init(characters: [String]) {
+        self = characters.map { .character($0) }
+    }
+    
+    /// Get a matching leading character margin action.
+    var leadingCharacterMarginAction: KeyboardAction {
+        characterMarginAction(for: first { $0.isInputAction })
+    }
+
+    /// Get a matching character margin action.
+    var trailingCharacterMarginAction: KeyboardAction {
+        characterMarginAction(for: last { $0.isInputAction })
+    }
+
+    /// Get a matching character margin action for an action.
+    func characterMarginAction(for action: KeyboardAction?) -> KeyboardAction {
+        guard let action = action else { return .none }
+        switch action {
+        case .character(let char): return .characterMargin(char)
+        default: return .none
+        }
+    }
+}
+
+public extension KeyboardAction.Rows {
+    
+    /// Map a string array array to keyboard action rows.
+    init(characters: [[String]]) {
+        self = characters.map { KeyboardAction.Row(characters: $0) }
+    }
+}